// src/app/models/auth-user.model.ts
export class AuthUser {
  success: number;
  message: number;
  id: number;
  name: string;
  first_name: string;
  middle_name: string;
  last_name: string;
  email: string;
  gender: string;
  roles: string[];
  roleId: string[];
  roleLabel: string[];
  permissions: string[];
  permissionLabel: string[];
  ability_id: number;
  dob: Date;
  mobile: number;
  photo: string;
  is_departmental: number;
  is_pwd_engineer: any;
  department_name: string;
  department_short_name: string;
  designation_id: number;
  department_id: number;
  circleId: number;
  circleName: string;
  divisionId: number;
  divnName: string;
  subDivisionId: number;
  subDivnName: string;
  sectionId: number;
  designation_name: string;
  designation_alias: string;
  email_verified_at: Date;
  created_at: Date;
  updated_at: Date;
  avatar_updated_at: Date;
  profile_updated_at: Date;
  remarks: string;
  inforce: number;
  display: number;
  distId: number;
  stackId: number;
  rlyId: number;
  officeId: number;
  userPosting: string;
  pwdProRemarks: string;
}